refactor(agents): split auth-profile mutation lineage out of runtime-snapshots
Split the max-lines boundary by moving persisted mutation lineage out of runtime snapshots; behavior is unchanged.
